requirements.txt
requests>=2.31.0
anthropic>=0.3.0
loguru>=0.7.0
python-dotenv>=1.0.0
notion-client>=2.0.0
pypdf2>=3.0.0